update routes to show DataVisSearchContainer, update reducer_data_vis to send mock payload, create DataVisSearchContainer for searching query, update DataVisContainer to accept only state.
